Upgrade Django and DRF. Fix tests. Fix serializers.
* Remove rnasync and client.
* Fix setup and MANIFEST.in to include static files and templates
* Add patch http method override middleware
* Fix serializer to include ID field.
