+++ .travis.yml
@@ -1,9 +1,9 @@
+sudo: false
 language: python
 python:
-  - "2.6"
   - "2.7"
 env:
-  - DJANGO_SETTINGS_MODULE=test_app.settings
+  - DJANGO_SETTINGS_MODULE=tests.settings
 before_script:
   - flake8 rna
 script: make cover
