+++ requirements/api.txt
@@ -1,3 +1,3 @@
 -r common.txt
 
-djangorestframework==2.3.7
+djangorestframework==3.3.2
