+++ requirements/common.txt
@@ -1,4 +1,3 @@
-South==0.8.2
-Django==1.4.9
+Django==1.8.9
 django-extensions==1.2.0
 django-pagedown==0.0.5
