+++ rna/static/js/releasenotes.js
@@ -9,7 +9,10 @@
   var notesApiUrl = releaseApiUrl + 'notes/';
 
   function authPost(url, data, callback, patchOverride) {
-    var headers = {'Content-Type': 'application/json'};
+    var headers = {
+      'Content-Type': 'application/json',
+      'X-CSRFToken': getCookie('csrftoken')
+    };
     if(patchOverride) {
       headers['X-HTTP-Method-Override'] = 'PATCH';
     }
@@ -54,6 +57,22 @@
 
   function bugUrl(bug) {
     return 'https://bugzilla.mozilla.org/show_bug.cgi?id=' + bug;
+  }
+
+  function getCookie(name) {
+    var cookieValue = null;
+    if (document.cookie && document.cookie != '') {
+      var cookies = document.cookie.split(';');
+      for (var i = 0; i < cookies.length; i++) {
+        var cookie = jQuery.trim(cookies[i]);
+        // Does this cookie string begin with the name we want?
+        if (cookie.substring(0, name.length + 1) == (name + '=')) {
+          cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+          break;
+        }
+      }
+    }
+    return cookieValue;
   }
 
   var BugLink = React.createClass({displayName: 'BugLink',
