+++ rna/urls.py
@@ -1,7 +1,8 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
-from django.conf.urls import patterns, url
+from django.conf.urls import url, include
+
 from rest_framework import routers
 
 from . import views
@@ -11,7 +12,8 @@ router = routers.DefaultRouter()
 router.register('notes', views.NoteViewSet)
 router.register('releases', views.ReleaseViewSet)
 
-urlpatterns = router.urls + patterns(
-    '',
+urlpatterns = [
+    url(r'^', include(router.urls)),
     url(r'^releases/(?P<pk>\d+)/notes/$', views.NestedNoteView.as_view()),
-    url(r'^auth_token/$', views.auth_token))
+    url(r'^auth_token/$', views.auth_token),
+]
