+++ tests/settings.py
@@ -4,7 +4,7 @@ TEMPLATE_DEBUG = DEBUG
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': 'test.db',
+        'NAME': ':memory:',
     }
 }
 
@@ -12,7 +12,7 @@ USE_TZ = False
 SITE_ID = 1
 SECRET_KEY = 'gxxg@@juj%4=-jr5ohv3cdj6)v6p2j5e3q91naw#m&amp;&amp;dgzq-zh'
 
-ROOT_URLCONF = 'test_app.urls'
+ROOT_URLCONF = 'tests.urls'
 STATIC_URL = '/static/'
 
 INSTALLED_APPS = (
@@ -24,8 +24,7 @@ INSTALLED_APPS = (
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django_extensions',
-    'south',
-    'django_nose',  # must come after south b/c south overrides test runner
+    'django_nose',
     'pagedown',
     'rest_framework',
     'rest_framework.authtoken',
@@ -34,13 +33,12 @@ INSTALLED_APPS = (
 
 TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
 NOSE_ARGS = ('--nocapture', )
-SOUTH_TESTS_MIGRATE = False
 
 REST_FRAMEWORK = {
     # Use hyperlinked styles by default.
     # Only used if the `serializer_class` attribute is not set on a view.
     'DEFAULT_MODEL_SERIALIZER_CLASS': (
-        'rna.rna.serializers.HyperlinkedModelSerializerWithPkField'),
+        'rna.serializers.HyperlinkedModelSerializerWithPkField'),
 
     # Use Django's standard `django.contrib.auth` permissions,
     # or allow read-only access for unauthenticated users.
@@ -52,7 +50,7 @@ REST_FRAMEWORK = {
         'rest_framework.authentication.TokenAuthentication',
     ),
 
-    'DEFAULT_FILTER_BACKENDS': ('rna.rna.filters.TimestampedFilterBackend',)
+    'DEFAULT_FILTER_BACKENDS': ('rna.filters.TimestampedFilterBackend',)
 }
 
 RNA = {'BASE_URL': 'https://nucleus.mozilla.org/rna/'}
